requirements.txt
# SPE Engine - Core Dependencies
# ================================

# Cryptographic signing (optional - system works without it)
pynacl>=1.5.0

# No other dependencies required.
# SPE Engine uses only Python standard library modules:
#   hashlib, json, sqlite3, time, zipfile, tempfile, pathlib, dataclasses
